+++ test/views/text.spec.js
@@ -63,7 +63,7 @@ function verifyTextViewContents(textView, data, format) {
             });
             break;
         default:
-            throw "verifyTextViewContents only supports json and raw formats";
+            throw 'verifyTextViewContents only supports json and raw formats';
     }
 }
 
@@ -311,17 +311,17 @@ describe('TextView Sink View', function () {
             verifyTextViewContents(chart, [pt1,pt2], 'raw');
         });
 
-        it("doesn't complain about timeless points", function() {
+        it('doesn\'t complain about timeless points', function() {
             var chart = new TextView({});
             chart.setDimensions(null, 200, 200);
 
             chart.consume([
                 {
-                    category: "A",
+                    category: 'A',
                     value: 1
                 },
                 {
-                    category: "B",
+                    category: 'B',
                     value: 1
                 }
             ]);