+++ test/views/timechart.spec.js
@@ -40,14 +40,14 @@ describe('Time Chart Sink View', function () {
                 });
 
                 chart.consume([
-                    { time : test_date, host : 'host1', pop : 'pop1', country : "Canada", value : 10 }
+                    { time : test_date, host : 'host1', pop : 'pop1', country : 'Canada', value : 10 }
                 ]);
 
                 chart.chart.series['0'].label.should.equal('country: Canada');
 
                 chart.consume([
-                    { time : test_date, host : 'host1', pop : 'pop2', country : "Canada", value : 10 },
-                    { time : test_date, host : 'host2', pop : 'pop2', country : "Canada", value : 10 }
+                    { time : test_date, host : 'host1', pop : 'pop2', country : 'Canada', value : 10 },
+                    { time : test_date, host : 'host2', pop : 'pop2', country : 'Canada', value : 10 }
                 ]);
 
                 chart.chart.series['0'].label.should.equal('host: host1, pop: pop1');
@@ -55,7 +55,7 @@ describe('Time Chart Sink View', function () {
                 chart.chart.series['2'].label.should.equal('host: host2, pop: pop2');
 
                 chart.consume([
-                    { time : test_date, host : 'host1', pop : 'pop2', country : "Portugal", value : 10 }
+                    { time : test_date, host : 'host1', pop : 'pop2', country : 'Portugal', value : 10 }
                 ]);
 
                 chart.chart.series['0'].label.should.equal('country: Canada, host: host1, pop: pop1');
@@ -87,12 +87,34 @@ describe('Time Chart Sink View', function () {
                 });
 
                 chart.consume([
-                    { time : test_date, value : 10, value2 : 20, host : "host1" },
-                    { time : test_date, value : 11, value2 : 30, host : "host2" }
+                    { time : test_date, value : 10, value2 : 20, host : 'host1' },
+                    { time : test_date, value : 11, value2 : 30, host : 'host2' }
                 ]);
 
                 chart.chart.series['0'].label.should.equal('host: host1');
                 chart.chart.series['1'].label.should.equal('host: host2');
+            });
+
+            it('null fields should be ignored for determining series', function() {
+                var chart = new TimeChartView({
+                    juttleEnv : juttleEnv,
+                    params : {
+                        valueField : 'value'
+                    }
+                });
+
+                // The second two points should fall into the two series of the first two points
+                // despite having an extra field with a `null` value.
+                chart.consume([
+                    {time: test_date, name: 'metric1', host: 'A', value: 1},
+                    {time: test_date, name: 'metric2', host: 'A', value: 1},
+                    {time: new Date(test_date.getTime() + 1000), name: 'metric1', host: 'A', value: 2, numCores: null},
+                    {time: new Date(test_date.getTime() + 1000), name: 'metric2', host: 'A', value: 2, numCores: null}
+                ]);
+
+                chart.chart.series['0'].label.should.equal('name: metric1');
+                chart.chart.series['1'].label.should.equal('name: metric2');
+                Object.keys(chart.chart.series).length.should.equal(2);
             });
         });
 
@@ -103,8 +125,8 @@ describe('Time Chart Sink View', function () {
                 });
 
                 chart.consume([
-                    { time : test_date, value : 10, common: "test", not_common: "shown" },
-                    { time : test_date, value : 10, common: "test1" }
+                    { time : test_date, value : 10, common: 'test', not_common: 'shown' },
+                    { time : test_date, value : 10, common: 'test1' }
                 ]);
 
                 chart.chart.series['0'].label.should.equal('common: test, not_common: shown');
@@ -117,8 +139,8 @@ describe('Time Chart Sink View', function () {
                 });
 
                 chart.consume([
-                    { time : test_date, value : 10, common: "test", some_field: "shown" },
-                    { time : test_date, value : 10, common: "test"}
+                    { time : test_date, value : 10, common: 'test', some_field: 'shown' },
+                    { time : test_date, value : 10, common: 'test'}
                 ]);
 
                 chart.chart.series['0'].label.should.equal('some_field: shown');
@@ -274,7 +296,7 @@ describe('Time Chart Sink View', function () {
             viewTestUtils.verifyValidationError({
                 viewConstructor : TimeChartView,
                 params : {
-                    interval : "someString"
+                    interval : 'someString'
                 },
                 errorPath : 'interval',
                 error : {